# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DJ-Blog API文档",
        "title": "DJ-Blog API文档",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/user/login": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录请求",
                        "name": "UserLoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Status"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "请求参数格式错误",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "402": {
                        "description": "请求参数不符合要求",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            }
        },
...
        "/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册请求",
                        "name": "userRegisterRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Status"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "请求参数格式错误",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "402": {
                        "description": "请求参数不符合要求",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CategoryModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.TagModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "request.UserLoginReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "required;min:6;max:20"
                },
                "username": {
                    "type": "string",
                    "example": "required;min:2;max:20"
                }
            }
        },
        "request.UserRegisterReq": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "example": "required;min:6;max:20"
                },
                "username": {
                    "description": "example 是 swagger 文档中的示例的关键字，用于生成 swagger 文档。\n这里的值并没有使用一个真的示例，而是用了这个数据的限制作为示例。\nTODO: 利用这个example值，做自动化的测试，限制条件之间用分号分隔，如：min:2,max:20",
                    "type": "string",
                    "example": "required;min:2;max:20"
                }
            }
        },
        "response.Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "service.Article": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CategoryModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imgUrl": {
                    "type": "string"
                },
                "originContent": {
                    "type": "string"
                },
                "parseContent": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TagModel"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "service.User": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Article"
                    }
                },
                "avatarUrl": {
                    "type": "string",
                    "example": "https://www.baidu.com/img/PCtm_d9c8750bed0b3c7d089fa7d55720d6cf.png"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2021-01-01 00:00:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2021-01-01 00:00:00"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        }
    }
}